requirements.txt
discord.py
beautifulsoup4
dotenv
requests
urllib.request